# PERF: use array buffer methods for plot curve line/symbol data retrieval