## Changes

* Use `tao.plot_line()` with `x_or_y` to take advantage of the `tao_c_interface_com%c_real` array buffer.
* Set `plot_page%delete_overlapping_plots` to `False` to ensure plot regions don't affect one another

## Background

The current `master` branch uses `tao.plot_line()` to retrieve both X- and Y- positions of the given plot curve.
This offers a significant speed-up for larger lattices. 

For example, making plots of a lattice with a few thousand elements may get a considerable speed-up with pytao plotting (i.e., matplotlib or bokeh) after this PR is merged:

`master` under the hood uses this:
```
In [1]: %timeit tao.plot_line("r13", "g", "a", x_or_y="")
44 ms ± 932 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

This PR uses this:
```
In [14]: %timeit tao.plot_line("r13", "g", "a", x_or_y="x")
983 μs ± 52.6 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [17]: %timeit tao.plot_line("r13", "g", "a", x_or_y="y")
1.07 ms ± 28.7 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

For a single curve, this PR would accelerate plots (and interactive updates of those plots) by ~22x (`=44/(1+0.9) ms`).
This speed-up applies to every curve displayed, so applications with many curves will see a much more obvious improvement.